Updating CLI.
