+++ Command_Line_Interface/modules/__init__.py
@@ -1,0 +1,2 @@
+__author__ = "Jarrod N. Bakker"
+__status__ = "development"
